refactor: 全体設計の見直し
