+++ src/index.js
@@ -1,0 +1,4 @@
+import { TetrisApp } from './components/TetrisApp.js';
+
+// アプリケーションのインスタンスを作成
+new TetrisApp();
